Add more comments in late fusion, remove general helper functions from AdvCoperceptionIntermediateFusionAttack
