+++ opencda/core/attack/advcp/intermediate_fusion_attack.py
@@ -1417,12 +1417,12 @@ class AdvCoperceptionIntermediateFusionAttack:
         # single row.
         probabilities = torch.sigmoid(output_dict["ego"]["psm"].permute(0, 2, 3, 1)).reshape(-1)
         proposals = dataset.post_processor.delta_to_boxes3d(output_dict["ego"]["rm"], batch_data["ego"]["anchor_box"])[0]
-        proposals_lwh = cls._model_boxes_to_lwh(proposals, dataset)
-        target_boxes_lwh = cls._model_boxes_to_lwh(target_boxes, dataset)
+        proposals_lwh = AdvCPAttackHelper.model_boxes_to_lwh(proposals, dataset)
+        target_boxes_lwh = AdvCPAttackHelper.model_boxes_to_lwh(target_boxes, dataset)
 
         loss_terms: list[torch.Tensor] = []
         for target_box in target_boxes_lwh:
-            iou_weights = cls._compute_iou_weights(proposals_lwh, target_box)
+            iou_weights = AdvCPAttackHelper.compute_iou_weights(proposals_lwh, target_box)
             # 0.01 threshold filters out proposals so distant from the
             # target that they would contribute negligible gradient.
             box_mask = iou_weights >= 0.01
@@ -1461,12 +1461,12 @@ class AdvCoperceptionIntermediateFusionAttack:
         """
         probabilities = torch.sigmoid(output_dict["ego"]["psm"].permute(0, 2, 3, 1)).reshape(-1)
         proposals = dataset.post_processor.delta_to_boxes3d(output_dict["ego"]["rm"], batch_data["ego"]["anchor_box"])[0]
-        proposals_lwh = cls._model_boxes_to_lwh(proposals, dataset)
-        target_boxes_lwh = cls._model_boxes_to_lwh(target_boxes, dataset)
+        proposals_lwh = AdvCPAttackHelper.model_boxes_to_lwh(proposals, dataset)
+        target_boxes_lwh = AdvCPAttackHelper.model_boxes_to_lwh(target_boxes, dataset)
 
         loss_terms: list[torch.Tensor] = []
         for target_box in target_boxes_lwh:
-            iou_weights = cls._compute_iou_weights(proposals_lwh, target_box)
+            iou_weights = AdvCPAttackHelper.compute_iou_weights(proposals_lwh, target_box)
             box_mask = iou_weights >= 0.01
             if torch.any(box_mask):
                 # Clamp away from zero to keep log() finite.
@@ -1476,270 +1476,3 @@ class AdvCoperceptionIntermediateFusionAttack:
         if not loss_terms:
             return probabilities.sum() * 0.0
         return torch.stack(loss_terms).sum()
-
-    @staticmethod
-    def _model_boxes_to_lwh(boxes: torch.Tensor, dataset: Any) -> torch.Tensor:
-        """
-        Reorder model-format boxes into the canonical ``lwh`` order.
-
-        OpenCOOD models can be configured with either ``"lwh"`` or
-        ``"hwl"`` axis ordering for the box dimensions. The IoU
-        computation expects ``"lwh"``; this helper rearranges the
-        columns when needed and returns the original tensor when the
-        ordering is already correct.
-
-        Raises
-        ------
-        NotImplementedError
-            If the dataset's order is not one of ``"hwl"`` or ``"lwh"``.
-        """
-        if boxes.numel() == 0:
-            return boxes
-        order = dataset.post_processor.params.get("order", "hwl")
-        if order == "hwl":
-            return boxes[:, [0, 1, 2, 5, 4, 3, 6]]
-        if order == "lwh":
-            return boxes
-        raise NotImplementedError(f"Unsupported box order for AdvCP intermediate spoofing: {order}")
-
-    @staticmethod
-    def _compute_iou_weights(proposals_lwh: torch.Tensor, target_box_lwh: torch.Tensor) -> torch.Tensor:
-        """
-        Compute 2D IoU between every proposal and a single target box.
-
-        Both the proposals and the target are projected to their 2D
-        oriented bounding boxes (XY corners only); the intersection
-        area is computed analytically by
-        :meth:`_oriented_box_intersection_2d` and the union derived
-        from individual areas.
-
-        Returns
-        -------
-        torch.Tensor
-            ``(num_proposals,)`` IoU values in ``[0, 1]``.
-        """
-        if proposals_lwh.shape[0] == 0:
-            return torch.zeros((0,), dtype=target_box_lwh.dtype, device=target_box_lwh.device)
-
-        # Broadcast the target across all proposals; both tensors then
-        # share the leading "proposal" dimension and can be processed
-        # in one batched call.
-        repeated_target_boxes = target_box_lwh.unsqueeze(0).expand(proposals_lwh.shape[0], -1)
-        proposal_corners = box_utils.boxes_to_corners2d(proposals_lwh, order="lwh")[:, :, :2]
-        target_corners = box_utils.boxes_to_corners2d(repeated_target_boxes, order="lwh")[:, :, :2]
-        intersection_area = AdvCoperceptionIntermediateFusionAttack._oriented_box_intersection_2d(
-            proposal_corners,
-            target_corners,
-        )
-        # 2D area = length * width (the L and W columns of the
-        # ordered box, indices 3 and 4).
-        proposal_area = proposals_lwh[:, 3] * proposals_lwh[:, 4]
-        target_area = repeated_target_boxes[:, 3] * repeated_target_boxes[:, 4]
-        union_area = torch.clamp(proposal_area + target_area - intersection_area, min=1e-6)
-        return torch.clamp(intersection_area / union_area, min=0.0, max=1.0)
-
-    @classmethod
-    def _oriented_box_intersection_2d(
-        cls,
-        corners1: torch.Tensor,
-        corners2: torch.Tensor,
-    ) -> torch.Tensor:
-        """
-        Differentiable 2D oriented-box intersection area, batched.
-
-        Implements the standard polygon-clipping recipe:
-
-        1. The intersection polygon is the convex hull of three vertex
-           sets:
-
-           - corners of box 1 lying inside box 2,
-           - corners of box 2 lying inside box 1,
-           - edge-edge intersection points between the two boxes.
-
-        2. Sort the vertices of that convex set by angle around their
-           centroid and compute the polygon area via the shoelace
-           formula.
-
-        Both arguments must be batches of 4-corner XY arrays (shape
-        ``(B, 4, 2)``); the result is a length-``B`` tensor of
-        intersection areas. Differentiable everywhere except at the
-        zero-area boundary, which is good enough for gradient-based
-        attacks.
-        """
-        intersections, intersection_mask = cls._box_intersection_th(corners1, corners2)
-        corners1_in_corners2, corners2_in_corners1 = cls._box_in_box_th(corners1, corners2)
-        # Concatenate up to 24 candidate vertices per box pair: 4
-        # corners1 + 4 corners2 + up to 16 edge-edge intersections.
-        # The boolean mask retains only the ones that actually belong
-        # to the intersection polygon.
-        vertices = torch.cat(
-            [
-                corners1,
-                corners2,
-                intersections.reshape(corners1.shape[0], -1, 2),
-            ],
-            dim=1,
-        )
-        vertex_mask = torch.cat(
-            [
-                corners1_in_corners2,
-                corners2_in_corners1,
-                intersection_mask.reshape(corners1.shape[0], -1),
-            ],
-            dim=1,
-        )
-        return cls._calculate_polygon_area(vertices, vertex_mask)
-
-    @staticmethod
-    def _box_intersection_th(
-        corners1: torch.Tensor,
-        corners2: torch.Tensor,
-    ) -> tuple[torch.Tensor, torch.Tensor]:
-        """
-        Compute every edge-edge intersection between two batched
-        quadrilaterals.
-
-        For each box pair, builds the four directed edges
-        ``(corners[i], corners[i+1])``, then computes the line-line
-        intersection between every pair of edges (4 x 4 = 16 per pair).
-        The mask flags valid intersections (parameters ``t`` and ``u``
-        both in ``(0, 1)``); invalid entries are zeroed in the
-        intersection tensor so downstream code can ignore them via the
-        mask.
-
-        Returns
-        -------
-        tuple
-            ``(intersections, mask)`` of shapes
-            ``(B, 4, 4, 2)`` and ``(B, 4, 4)``.
-        """
-        epsilon = 1e-8
-        lines1 = torch.cat([corners1, corners1[:, [1, 2, 3, 0], :]], dim=2)
-        lines2 = torch.cat([corners2, corners2[:, [1, 2, 3, 0], :]], dim=2)
-
-        lines1_expanded = lines1.unsqueeze(2).repeat(1, 1, 4, 1)
-        lines2_expanded = lines2.unsqueeze(1).repeat(1, 4, 1, 1)
-        x1 = lines1_expanded[..., 0]
-        y1 = lines1_expanded[..., 1]
-        x2 = lines1_expanded[..., 2]
-        y2 = lines1_expanded[..., 3]
-        x3 = lines2_expanded[..., 0]
-        y3 = lines2_expanded[..., 1]
-        x4 = lines2_expanded[..., 2]
-        y4 = lines2_expanded[..., 3]
-
-        denominator = (x1 - x2) * (y3 - y4) - (y1 - y2) * (x3 - x4)
-        denominator_t = (x1 - x3) * (y3 - y4) - (y1 - y3) * (x3 - x4)
-        denominator_u = (x1 - x2) * (y1 - y3) - (y1 - y2) * (x1 - x3)
-
-        t = denominator_t / torch.where(denominator == 0, torch.ones_like(denominator), denominator)
-        u = -denominator_u / torch.where(denominator == 0, torch.ones_like(denominator), denominator)
-        t = torch.where(denominator == 0, torch.full_like(t, -1.0), t)
-        u = torch.where(denominator == 0, torch.full_like(u, -1.0), u)
-
-        mask_t = (t > 0) & (t < 1)
-        mask_u = (u > 0) & (u < 1)
-        intersection_mask = mask_t & mask_u
-
-        stable_t = denominator_t / (denominator + epsilon)
-        intersections = torch.stack(
-            [
-                x1 + stable_t * (x2 - x1),
-                y1 + stable_t * (y2 - y1),
-            ],
-            dim=-1,
-        )
-        intersections = intersections * intersection_mask.unsqueeze(-1).to(intersections.dtype)
-        return intersections, intersection_mask
-
-    @classmethod
-    def _box_in_box_th(
-        cls,
-        corners1: torch.Tensor,
-        corners2: torch.Tensor,
-    ) -> tuple[torch.Tensor, torch.Tensor]:
-        """
-        Test corner containment in both directions.
-
-        Returns a pair of boolean tensors: which of ``corners1`` lie
-        inside ``corners2`` and vice versa.
-        """
-        return cls._box1_in_box2(corners1, corners2), cls._box1_in_box2(corners2, corners1)
-
-    @staticmethod
-    def _box1_in_box2(corners1: torch.Tensor, corners2: torch.Tensor) -> torch.Tensor:
-        """
-        Boolean mask: which corners of box 1 lie inside box 2.
-
-        Treats box 2 as a parallelogram defined by its corner ``A``
-        and the edge vectors ``AB`` and ``AD``. A point ``M`` lies
-        inside iff its parametric coordinates along both edges are in
-        ``[0, 1]``:
-
-        ``s = (AM . AB) / (AB . AB), t = (AM . AD) / (AD . AD)``
-
-        The check uses a small slack (``+/-1e-6``) to keep boundary
-        points inside.
-        """
-        corner_a = corners2[:, 0:1, :]
-        corner_b = corners2[:, 1:2, :]
-        corner_d = corners2[:, 3:4, :]
-        vector_ab = corner_b - corner_a
-        vector_am = corners1 - corner_a
-        vector_ad = corner_d - corner_a
-        projection_ab = torch.sum(vector_ab * vector_am, dim=-1)
-        norm_ab = torch.sum(vector_ab * vector_ab, dim=-1)
-        projection_ad = torch.sum(vector_ad * vector_am, dim=-1)
-        norm_ad = torch.sum(vector_ad * vector_ad, dim=-1)
-
-        condition_ab = (projection_ab / norm_ab > -1e-6) & (projection_ab / norm_ab < 1.0 + 1e-6)
-        condition_ad = (projection_ad / norm_ad > -1e-6) & (projection_ad / norm_ad < 1.0 + 1e-6)
-        return condition_ab & condition_ad
-
-    @staticmethod
-    def _calculate_polygon_area(vertices: torch.Tensor, vertex_mask: torch.Tensor) -> torch.Tensor:
-        """
-        Shoelace area of a batched polygon with masked vertices.
-
-        For each batch element:
-
-        1. Compute the centroid of valid vertices (their mean).
-        2. Sort all vertices by their angle around that centroid;
-           invalid vertices are pushed to the end via a ``+inf``
-           sentinel angle so they end up unused.
-        3. Apply the shoelace formula on the first ``num_valid``
-           vertices.
-
-        Polygons with fewer than three valid vertices have area zero
-        (no actual intersection).
-
-        Returns
-        -------
-        torch.Tensor
-            ``(B,)`` polygon areas. Differentiable in vertex
-            coordinates.
-        """
-        num_valid = vertex_mask.to(torch.int32).sum(dim=1)
-        safe_num_valid = torch.clamp(num_valid, min=1).to(vertices.dtype)
-        centroid = (vertices * vertex_mask.unsqueeze(-1).to(vertices.dtype)).sum(dim=1) / safe_num_valid.unsqueeze(-1)
-        normalized_vertices = vertices - centroid.unsqueeze(1)
-        angles = torch.atan2(normalized_vertices[..., 1], normalized_vertices[..., 0])
-        invalid_angle = torch.full_like(angles, float("inf"))
-        sorted_indices = torch.argsort(torch.where(vertex_mask, angles, invalid_angle), dim=1)
-
-        gather_indices = sorted_indices.unsqueeze(-1).expand(-1, -1, 2)
-        sorted_vertices = torch.gather(vertices, 1, gather_indices)
-        max_vertices = sorted_vertices.shape[1]
-        vertex_indices = torch.arange(max_vertices, device=vertices.device).unsqueeze(0).expand(sorted_vertices.shape[0], -1)
-        next_indices = torch.where(
-            vertex_indices + 1 < num_valid.unsqueeze(1),
-            vertex_indices + 1,
-            torch.zeros_like(vertex_indices),
-        )
-        valid_edge_mask = vertex_indices < num_valid.unsqueeze(1)
-        next_gather_indices = next_indices.unsqueeze(-1).expand(-1, -1, 2)
-        next_vertices = torch.gather(sorted_vertices, 1, next_gather_indices)
-
-        cross_products = sorted_vertices[..., 0] * next_vertices[..., 1] - sorted_vertices[..., 1] * next_vertices[..., 0]
-        polygon_area = torch.abs((cross_products * valid_edge_mask.to(sorted_vertices.dtype)).sum(dim=1)) / 2.0
-        return torch.where(num_valid >= 3, polygon_area, torch.zeros_like(polygon_area))
